// src/app/models/registro.ts
export interface registro {
  username: string;
  email: string;
  password: string;
  nombre: string;
  apellido: string;
  telefono: string;
  idMunicipio: number;
  idDocumento:number;
  numeroDocumento:number;
  rol: string;
  descripcion: string;
  idServicio: number;
  foto: string | null;
  cv: string | null;
}